requirements.txt
flask
flask_sqlalchemy
flask_login
torch
torchvision
pillow
pandas
weasyprint
